+++ Sources/BraintreeUIKit/Components/BTUIKFormField.m
@@ -46,6 +46,7 @@
         self.formLabel.translatesAutoresizingMaskIntoConstraints = NO;
         self.formLabel.text = @"";
         self.formLabel.accessibilityElementsHidden = YES;
+        self.formLabel.numberOfLines = 0;
         [self addSubview:self.formLabel];
 
         [self.formLabel setContentHuggingPriority:UILayoutPriorityRequired forAxis:UILayoutConstraintAxisHorizontal];
@@ -78,26 +79,20 @@
     
     BOOL hasFormLabel = (self.formLabel.text.length > 0);
     
-    [self.layoutConstraints addObjectsFromArray:[NSLayoutConstraint constraintsWithVisualFormat:@"V:|[textField]|"
+    [self.layoutConstraints addObjectsFromArray:[NSLayoutConstraint constraintsWithVisualFormat:@"V:|-(PADDING)-[textField]-(PADDING)-|"
                                                                  options:0
                                                                  metrics:metrics
                                                                    views:viewBindings]];
 
-    [self.layoutConstraints addObjectsFromArray:[NSLayoutConstraint constraintsWithVisualFormat:@"V:|[formLabel]|"
+    [self.layoutConstraints addObjectsFromArray:[NSLayoutConstraint constraintsWithVisualFormat:@"V:|-(PADDING)-[formLabel]-(PADDING)-|"
                                                                  options:0
                                                                  metrics:metrics
                                                                    views:viewBindings]];
-    if (hasFormLabel) {
-        [self.layoutConstraints addObjectsFromArray:[NSLayoutConstraint constraintsWithVisualFormat:@"H:|-(PADDING)-[formLabel(<=0@1)]-[textField]"
-                                                                                            options:0
-                                                                                            metrics:metrics
-                                                                                              views:viewBindings]];
-    } else {
-        [self.layoutConstraints addObjectsFromArray:[NSLayoutConstraint constraintsWithVisualFormat:@"H:|-(PADDING)-[textField]"
-                                                                                            options:0
-                                                                                            metrics:metrics
-                                                                                              views:viewBindings]];
-    }
+    
+    [self.layoutConstraints addObjectsFromArray:[NSLayoutConstraint constraintsWithVisualFormat:@"H:|-(PADDING)-[formLabel(<=0@1)]-[textField]"
+                                                                                        options:0
+                                                                                        metrics:metrics
+                                                                                          views:viewBindings]];
     
     if (self.accessoryView && !self.accessoryView.hidden) {
         [self.layoutConstraints addObjectsFromArray:@[[NSLayoutConstraint constraintWithItem:self.accessoryView
@@ -107,8 +102,6 @@
                                                                                  attribute:NSLayoutAttributeCenterY
                                                                                 multiplier:1.0f
                                                                                   constant:0.0f]]];
-        
-      ;
         
         [self.layoutConstraints addObjectsFromArray:[NSLayoutConstraint constraintsWithVisualFormat:@"H:[textField]-[accessoryView]-(PADDING)-|"
                                                                                             options:0
@@ -211,7 +204,7 @@
     
     NSMutableAttributedString *mutableText = [[NSMutableAttributedString alloc] initWithAttributedString:self.textField.attributedText];
     [mutableText addAttributes:@{NSForegroundColorAttributeName: textColor,
-                                 NSFontAttributeName:[[BTUIKAppearance sharedInstance].font fontWithSize:UIFont.labelFontSize]}
+                                 NSFontAttributeName:[BTUIKAppearance sharedInstance].bodyFont}
                          range:NSMakeRange(0, mutableText.length)];
     
     UITextRange *currentRange = self.textField.selectedTextRange;