+++ Sources/BraintreeUIKit/Helpers/BTUIKAppearance.m
@@ -6,13 +6,6 @@
 #import <BraintreeUIKit/UIColor+BTUIK.h>
 #endif
 
-@interface BTUIKAppearance ()
-
-@property (nonatomic, strong) UIFont *font;
-@property (nonatomic, strong) UIFont *boldFont;
-
-@end
-
 @implementation BTUIKAppearance
 
 static BTUIKAppearance *sharedTheme;
@@ -26,60 +19,43 @@ static BTUIKAppearance *sharedTheme;
     return sharedTheme;
 }
 
-- (void)setFontFamily:(NSString *)fontFamily {
-    _fontFamily = fontFamily;
-    if (fontFamily != nil) {
-        _font = [UIFont fontWithName:fontFamily size:10.0];
-    } else {
-        _font = [UIFont systemFontOfSize:10.0];
-    }
-}
-
-- (void)setBoldFontFamily:(NSString *)boldFontFamily {
-    _boldFontFamily = boldFontFamily;
-    if (boldFontFamily != nil) {
-        _boldFont = [UIFont fontWithName:boldFontFamily size:10.0];
-    } else {
-        _boldFont = [UIFont boldSystemFontOfSize:10.0];
-    }
-}
-
 - (UIColor *)highlightedTintColor {
     return [sharedTheme.tintColor colorWithAlphaComponent:0.4];
 }
 
 + (void)styleLabelPrimary:(UILabel *)label {
-    label.font = [[BTUIKAppearance sharedInstance].font fontWithSize:UIFont.labelFontSize];
+    label.font = [BTUIKAppearance sharedInstance].bodyFont;
+    label.adjustsFontForContentSizeCategory = YES;
     label.textColor = [BTUIKAppearance sharedInstance].primaryTextColor;
 }
 
 + (void)styleLabelBoldPrimary:(UILabel *)label {
-    label.font = [[BTUIKAppearance sharedInstance].boldFont fontWithSize:UIFont.labelFontSize];
-    label.textColor = [BTUIKAppearance sharedInstance].primaryTextColor;
-}
-
-+ (void)styleSmallLabelBoldPrimary:(UILabel *)label {
-    label.font = [[BTUIKAppearance sharedInstance].boldFont fontWithSize:UIFont.smallSystemFontSize];
+    label.font = [BTUIKAppearance sharedInstance].headlineFont;
+    label.adjustsFontForContentSizeCategory = YES;
     label.textColor = [BTUIKAppearance sharedInstance].primaryTextColor;
 }
 
 + (void)styleSmallLabelPrimary:(UILabel *)label {
-    label.font = [[BTUIKAppearance sharedInstance].font fontWithSize:UIFont.smallSystemFontSize];
+    label.font = [BTUIKAppearance sharedInstance].captionFont;
+    label.adjustsFontForContentSizeCategory = YES;
     label.textColor = [BTUIKAppearance sharedInstance].primaryTextColor;
 }
 
 + (void)styleLabelSecondary:(UILabel *)label {
-    label.font = [[BTUIKAppearance sharedInstance].font fontWithSize:UIFont.smallSystemFontSize];
+    label.font = [BTUIKAppearance sharedInstance].captionFont;
+    label.adjustsFontForContentSizeCategory = YES;
     label.textColor = [BTUIKAppearance sharedInstance].secondaryTextColor;
 }
 
 + (void)styleLargeLabelSecondary:(UILabel *)label {
-    label.font = [[BTUIKAppearance sharedInstance].font fontWithSize:UIFont.labelFontSize];
+    label.font = [BTUIKAppearance sharedInstance].bodyFont;
+    label.adjustsFontForContentSizeCategory = YES;
     label.textColor = [BTUIKAppearance sharedInstance].secondaryTextColor;
 }
 
 + (void)styleSystemLabelSecondary:(UILabel *)label {
-    label.font = [[BTUIKAppearance sharedInstance].font fontWithSize:UIFont.systemFontSize];
+    label.font = [BTUIKAppearance sharedInstance].staticBodyFont;
+    label.adjustsFontForContentSizeCategory = YES;
     label.textColor = [BTUIKAppearance sharedInstance].secondaryTextColor;
 }
 
@@ -87,7 +63,7 @@ static BTUIKAppearance *sharedTheme;
     UILabel *tlabel = [[UILabel alloc] initWithFrame:CGRectMake(0,0, 200, 40)];
     tlabel.textAlignment = NSTextAlignmentCenter;
     tlabel.textColor = [BTUIKAppearance sharedInstance].navigationBarTitleTextColor;
-    tlabel.font = [[BTUIKAppearance sharedInstance].boldFont fontWithSize:UIFont.labelFontSize];
+    tlabel.font = [BTUIKAppearance sharedInstance].staticHeadlineFont;
     tlabel.backgroundColor = UIColor.clearColor;
     tlabel.adjustsFontSizeToFitWidth = YES;
     tlabel.numberOfLines = 2;
@@ -135,7 +111,6 @@ static BTUIKAppearance *sharedTheme;
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{
         sharedMetrics = @{@"HORIZONTAL_FORM_PADDING":@([BTUIKAppearance horizontalFormContentPadding]),
-                          @"FORM_CELL_HEIGHT":@([BTUIKAppearance formCellHeight]),
                           @"VERTICAL_FORM_SPACE":@([BTUIKAppearance verticalFormSpace]),
                           @"VERTICAL_FORM_SPACE_TIGHT":@([BTUIKAppearance verticalFormSpaceTight]),
                           @"VERTICAL_SECTION_SPACE":@([BTUIKAppearance verticalSectionSpace]),