+++ CHANGELOG.md
@@ -13,6 +13,7 @@
     * `BTPaymentSelectionViewController`
     * `BTVaultManagementViewController`
   * Add `BTDropInUICustomization`, which replaces `BTUIKAppearance` for customizing UI
+  * Enable dynamic type by default. This can be disabled with the `disableDynamicType` property on `BTDropInUICustomization`.
 
 ## 8.1.2 (2020-11-30)